requirements.txt
boto3==1.33.13
Flask==2.2.3
flask-cors~=5.0.0
Pillow==9.5.0
PyMuPDF==1.23.5
requests~=2.32.3
waitress==2.1.2
Werkzeug==2.2.2
pypdf==4.0.1
botocore~=1.35.55